latihan minggu ke 2 hari ke empat

+++ Novice/02-03/kasus/src/main/java/kasus/concurrency2/ConsumerProducer.java
@@ -19,7 +19,7 @@ public class ConsumerProducer {
                     }
                 }
                 if (!queue.isEmpty()) {
-                    Integer integer = queue.poll();
+                    Integer integer = (Integer) queue.poll();
                     System.out.println("[" + Thread.currentThread().getName() + "]: " + integer);
                 }
             }
