+++ Novice/02-03/kasus/src/main/java/kasus/concurrency2/SynchronizedAndWait.java
@@ -15,7 +15,7 @@ public class SynchronizedAndWait {
                 } catch (InterruptedException e) {
                     e.printStackTrace();
                 }
-                retVal = queue.poll();
+                retVal = (Integer) queue.poll();
             }
         }
         return retVal;
